Understanding idx()
Commented out printing out the graph laplacian.  The matrix creates and prints, but something is wrong with the number of neighbors.  It always correctly identifies the right neighbors, but it also identifies vertices as neighbors that should not be.  This code will print out a list of pairs, each pair corresponding to a vertex and one of its supposed neighbors.